Additional cleanup and tweaks

+++ RAN_Intelligent_Controller/install_scripts/run_xapp_onboarder.sh
@@ -23,9 +23,16 @@ if ! dpkg -l | grep -q python3-venv; then
     sudo apt install -y python3-venv
 fi
 
-if [ -d "venv" ]; then
-    echo "Cleaning up previous virtual environment before creating a new one..."
-    sudo rm -rf venv
+# Check if the dmi_cli binary is already installed
+export CHART_REPO_URL="http://0.0.0.0:8090"
+if [[ $(sudo -E dms_cli health) == "True" ]]; then
+    echo "Health check was successful."
+    exit 0
+else
+    if [ -d "venv" ]; then
+        echo "Cleaning up previous virtual environment before creating a new one..."
+        sudo rm -rf venv
+    fi
 fi
 
 # Create a virtual environment and activate it
@@ -85,8 +92,6 @@ sudo ln -s "$DMS_CLI_PATH" /usr/local/bin/dms_cli
 echo "Symbolic link created for dms_cli at /usr/local/bin/dms_cli"
 
 cd ../../../../ # Main directory
-
-export CHART_REPO_URL="http://0.0.0.0:8090"
 
 if ! curl -s $CHART_REPO_URL > /dev/null; then
     echo "Server at http://0.0.0.0:8090 is not running. Attempting to start..."
